Replace CategoricalLSTMPolicy with Model
